+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Tous les changements notables de Ara sont documentés ici avec leur date, leur catégorie (nouvelle fonctionnalité, correction de bug ou autre changement) et leur pull request (PR) associée.
 
+## 08/02/2023
+
+### Corrections 🐛
+
+- Clarifie le statut de l’audit côté rapport ([#284](https://github.com/DISIC/Ara/pull/284))
+
 ## 03/02/2023
 
 ### Autres changements ⚙️